pyproject.toml
[project]
name = "robotics-nodes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "torchdiffeq>=0.2.5",
    "tqdm>=4.67.1",
]
